Improved file structure and added improved aimbot alive check
